requirements.txt
torch
transformers
scikit-learn
pandas
numpy
fastapi
uvicorn
s3fs
boto3
python-dotenv
PyYAML
